requirements.txt
# Flask フレームワーク
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0

# データベース
psycopg2-binary==2.9.9
alembic==1.13.1

# HTTP クライアント
requests==2.31.0

# データ処理
pandas==2.1.4
numpy==1.25.2

# Yahoo Finance
yfinance==0.2.18

# ユーティリティ
python-dotenv==1.0.0
gunicorn==21.2.0

# 開発・テスト用
pytest==7.4.3
pytest-flask==1.3.0
black==23.12.0
flake8==6.1.0